requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pydantic
python-multipart
python-dotenv
psycopg2-binary
uuid
pyjwt 